Fix printf format errors

+++ applis/eperftool/receiver.c
@@ -85,7 +85,7 @@ receive_and_decode (void)
 	OF_PRINT(("decoding_start=%lI64f\n", (double)tv0.QuadPart/(double)freq.QuadPart))
 #else
 	gettimeofday(&tv0, NULL);
-	OF_PRINT(("decoding_start=%ld.%d\n", tv0.tv_sec, tv0.tv_usec))
+	OF_PRINT(("decoding_start=%ld.%ld\n", tv0.tv_sec, (long)tv0.tv_usec))
 #endif
 
 	while ((new_symb_cb = get_next_symbol_received()) != NULL) {
@@ -275,9 +275,9 @@ receive_and_decode (void)
 #else
 		gettimeofday(&tv1, NULL);
 		timersub(&tv1, &tv0, &tv_delta);
-		OF_PRINT(("decoding_end=%ld.%d  decoding_time=%ld.%06d  nb_received_symbols=%d  inefficiency_ratio=%.6f\n",
-			tv1.tv_sec, tv1.tv_usec,
-			tv_delta.tv_sec, tv_delta.tv_usec,
+		OF_PRINT(("decoding_end=%ld.%ld  decoding_time=%ld.%06ld  nb_received_symbols=%d  inefficiency_ratio=%.6f\n",
+			tv1.tv_sec, (long)tv1.tv_usec,
+			tv_delta.tv_sec, (long)tv_delta.tv_usec,
 			tot_nb_recvd_symbols, (double)tot_nb_recvd_symbols/(double)tot_nb_source_symbols))
 #endif
 #ifdef CHECK_INTEGRITY
